♻️ REFACTOR: 공통 sidebar css 코드 통일
