--- src/components/layout/sidebar/Sidebar.css
@@ -1,240 +1,0 @@
-/* 'paybooc-Bold' 폰트 불러오기 */
-@font-face {
-  font-family: "paybooc-Bold";
-  src: url("https://fastly.jsdelivr.net/gh/projectnoonnu/noonfonts_20-07@1.0/paybooc-Bold.woff2")
-      format("woff2"),
-    url("https://fastly.jsdelivr.net/gh/projectnoonnu/noonfonts_20-07@1.0/paybooc-Bold.woff")
-      format("woff");
-  font-weight: normal;
-  font-style: normal;
-}
-
-/* 전체 바디 스타일 설정 */
-body {
-  font-family: "paybooc-Bold", sans-serif;
-  width: 100%;
-  height: 100vh;
-  margin: 0;
-}
-
-/* 사이드바 스타일 */
-.nav {
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  width: 15rem; /* 기본 사이드바 너비 */
-  padding: 2rem; /* 콘텐츠 주변 여백 */
-  box-sizing: border-box;
-  background-color: #121824; /* 사이드바 배경색 */
-  color: white;
-  opacity: 0; /* 처음에 보이지 않게 설정 */
-  visibility: hidden;
-  box-shadow: 2px 0 2px rgba(0, 0, 0, 0.5);
-  border-radius: 0.1rem;
-}
-
-/* 작은 화면에 맞춘 사이드바 스타일 */
-@media (max-width: 768px) {
-  .nav {
-    width: 12rem; /* 화면이 작아졌을 때 사이드바 너비 축소 */
-  }
-}
-
-/* 더 작은 화면에 맞춘 스타일 */
-@media (max-width: 576px) {
-  .nav {
-    width: 10rem; /* 화면이 더 작아졌을 때 사이드바 너비 축소 */
-    padding: 10px; /* 작은 화면에서 여백 축소 */
-  }
-}
-
-/* 사이드바 닫기 버튼 스타일 */
-.navClose {
-  position: absolute;
-  top: 7px; /* 70%로 조정 */
-  right: 0.7px; /* 70%로 조정 */
-  width: 28px; /* 40px의 70%로 조정 */
-  cursor: pointer;
-}
-
-/* 닫기 버튼의 선 스타일 */
-.navClose:before,
-.navClose:after {
-  content: "";
-  display: block;
-  width: 2.8px; /* 4px의 70%로 조정 */
-  height: 19.6px; /* 28px의 70%로 조정 */
-  background-color: #f0f4f8;
-  border-radius: 10px;
-  position: absolute;
-  transition: all 0.2s linear;
-}
-
-/* 닫기 버튼의 왼쪽 대각선 선 */
-.navClose:before {
-  transform: rotate(-45deg);
-  top: 0;
-  left: 12.6px; /* 18px의 70%로 조정 */
-}
-
-/* 닫기 버튼의 오른쪽 대각선 선 */
-.navClose:after {
-  transform: rotate(45deg);
-  top: 0;
-  right: 12.6px; /* 18px의 70%로 조정 */
-}
-
-.navHeader {
-  display: flex;
-  align-items: center;
-  justify-content: center; /* 왼쪽 정렬 */
-  margin-bottom: 3rem; /* 아래 여백 */
-  margin-top: 2.5rem;
-  margin-left: -1rem;
-}
-
-/* 사이드바 제목 스타일 */
-.sidebarTitle {
-  font-family: "paybooc-Bold", sans-serif;
-  font-size: 1.5rem; /* 제목 크기 */
-  font-weight: bold;
-  text-transform: uppercase; /* 대문자 변환 */
-  color: white;
-  transition: color 0.3s;
-}
-
-.logo {
-  width: 7rem; /* 로고의 폭을 설정 */
-  height: auto; /* 비율에 맞춰 높이 자동 조정 */
-  display: inline-block; /* 인라인 블록으로 설정하여 텍스트와 수평 정렬 */
-  vertical-align: middle; /* 중간 정렬 */
-  margin-right: -1rem; /* 텍스트와의 간격 조정 */
-  margin-left: -2.3rem;
-}
-
-@media (max-width: 576px) {
-  .logo {
-    width: 4rem; /* 작은 화면에서 로고 크기 축소 */
-  }
-}
-
-.navLinkIcon {
-  padding: 0.4rem;
-}
-
-/* 사이드바가 열렸을 때 */
-.navOffcanvas1.is-opened {
-  z-index: 100;
-  left: 0;
-  opacity: 1;
-  visibility: visible;
-}
-
-/* 메뉴 리스트 스타일 */
-.navList {
-  list-style-type: none;
-  padding: 0;
-  margin: 1rem 0 0;
-}
-
-/* 개별 메뉴 항목 스타일 */
-
-/* 메뉴 링크 스타일 */
-.navLink {
-  display: flex;
-  align-items: center;
-  color: #fff;
-  text-decoration: none;
-  position: relative;
-  padding: 10px 20px;
-  width: 11rem;
-  transition: none; /* 애니메이션 제거 */
-  border-top-left-radius: 20px;
-  border-bottom-left-radius: 20px;
-  box-shadow: inset 5px 0 5px -3px rgba(0, 0, 0, 0.4);
-}
-
-/* hover, active, focus 효과 모두 동일하게 적용 */
-aside a:hover,
-aside a:active,
-aside a:focus {
-  color: #121824;
-  background-color: #f0f4f8;
-  outline: none;
-  position: relative;
-  border-top-left-radius: 200px;
-  border-bottom-left-radius: 200px;
-}
-
-aside a:hover::after,
-aside a:active::after,
-aside a:focus::after {
-  content: "";
-  position: absolute;
-  background-color: transparent;
-  bottom: 100%;
-  right: 0;
-  height: 35px;
-  width: 37px;
-  border-bottom-right-radius: 100%;
-  box-shadow: 0 20px 0 0 #f0f4f8;
-}
-
-aside a:hover::before,
-aside a:active::before,
-aside a:focus::before {
-  content: "";
-  position: absolute;
-  background-color: transparent;
-  top: 49px;
-  right: 0;
-  height: 35px;
-  width: 37px;
-  border-top-right-radius: 100%;
-  box-shadow: 0 -20px 0 0 #f0f4f8;
-}
-
-/* 로그아웃 버튼 스타일 */
-.logoutSection {
-  display: flex;
-  justify-content: center;
-  position: absolute;
-  bottom: 4rem; /* 사이드바 하단에서의 거리 */
-  width: 100%; /* 가로 정렬 */
-  padding: 1rem 0;
-}
-
-.logoutBtn {
-  font-family: "paybooc-Bold", sans-serif;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 130px;
-  padding: 8px;
-  background-color: #fff;
-  color: #121824;
-  border: none;
-  border-radius: 10px;
-  cursor: pointer;
-  transition: background-color 0.3s;
-  margin-right: 4.5rem;
-}
-
-.logoutBtn:hover {
-  background-color: #f0f0f0;
-}
-
-.logoutIcon {
-  margin-right: 8px;
-}
-
-/* 푸터 스타일 */
-.sidebarFooter {
-  text-align: center;
-  color: #fff;
-  font-size: 0.8rem;
-  position: absolute;
-  bottom: 3rem; /* 사이드바 하단에서의 거리 */
-  right: 0.7rem;
-  width: 100%;
-}
