+++ src/pages/user/User.js
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import UserSidebar from "../../components/layout/sidebar/UserSidebar";
 import UserContent from "../../components/layout/content/user/UserContent";
+import "../../styles/Content.css";
 
 const User = () => {
   const [isNavOpen, setNavOpen] = useState(false);
